chatbot_assistant/requirements.txt
python-dotenv
requests
openai
gradio
pillow